# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RND TECHNOSOFT CRM API
  version: 1.0.0
  description: |
    Multi-tenant CRM API with three-tier architecture:
    super admin, tenant admin, and staff.
servers:
  - url: http://localhost:5000/api
...
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create tenant admin and tenant workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, company_name, email, password]
              properties:
                name: { type: string }
                company_name: { type: string }
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "201":
          description: Account created
...
  /superadmin/tenants:
    get:
      tags: [Superadmin]
      summary: List all tenants
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tenant list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT